[COLAB-2918]-[balloon]: move ErrorDecoder to utils project; let the FeignConfigs inherit from an default, abstract configuration class
