+++ microservices/services/members-service/src/main/java/org/xcolab/service/members/config/FeignConfig.java
@@ -1,18 +1,13 @@
 package org.xcolab.service.members.config;
 
-import feign.codec.ErrorDecoder;
 import org.springframework.cloud.netflix.feign.EnableFeignClients;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
-@Component
-@EnableFeignClients(basePackages = {"org.xcolab.client.tracking"})
-@Profile("!test")
-public class FeignConfig {
+import org.xcolab.util.autoconfigure.AbstractFeignConfig;
 
-    @Bean
-    public ErrorDecoder getErrorDecoder() {
-        return new XColabErrorDecoder();
-    }
+@Component
+@Profile("!test")
+@EnableFeignClients(basePackages = {"org.xcolab.client.tracking"})
+public class FeignConfig extends AbstractFeignConfig {
 }
