+++ util/xcolab-utils/src/main/java/org/xcolab/util/autoconfigure/XColabErrorDecoder.java
@@ -1,4 +1,4 @@
-package org.xcolab.service.members.config;
+package org.xcolab.util.autoconfigure;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.ObjectMapper;
